Push style on box section and remove my patience
Gods...the flexbox stuff is kicking me...I spent on thing portion of the
webiste more time that I wanted.

First, h2 was under flexbox, but it was a f headache. It was messy, so
displayed as block and aligned at the center.

Second, the images part...well, it was hard to make them perfect
but it looks great, + add
transition for those sick styling.

Then, some hover over image.

And flexbox...I lost all my nerves on that thing and still code looks
like a code monkey got himself free from the cage! If remove box section
flexbox property - everything going to be messed up! Than, going to box
classes itself - ok, I got it for some magical reason it did not want to
do it easy way and I have to be a shaman with a drum, jumping around
this section like a madlad. Whoever will be reading that - know that I
spent entire evening to just make it right.

There were different options to style them as they are sitting now, but
all of them were bad and ugly. Sure, code looks like a mess, but hey -
if it's working, it's not stupid.

Holly Jesus, what an experience.

Though, I learned a lot and read about different options and even tried
them.