Add function for loader
Well, my first time doing JS. I had to refer to Traversy Media to
get understanding on how to create such thing. Had to modify it a little
as main.style.display = 'block' wasn't an option for me and
was conflicting with my styling.
Changed it to initial and it seems like it's working as a charm.

So, first time, but went hard and now I have a little bit of knowledge
about such functions